Second commit

+++ Oline_Food_App/Oline_Food_App/src/main/java/com/km/onliefoodapp/service/FoodMenuService.java
@@ -1,6 +1,7 @@
 package com.km.onliefoodapp.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -12,7 +13,10 @@ import com.km.onliefoodapp.dao.UserDao;
 import com.km.onliefoodapp.entity.FoodMenu;
 import com.km.onliefoodapp.entity.Role;
 import com.km.onliefoodapp.entity.User;
+import com.km.onliefoodapp.exception.FoodMenuIdIsNotPresent;
 import com.km.onliefoodapp.exception.NoSuchDataFoundException;
+import com.km.onliefoodapp.exception.UserDataNotFoundInTheDatabase;
+import com.km.onliefoodapp.exception.UserRoleDoesNotMatch;
 import com.km.onliefoodapp.util.ResponseStructure;
 
 @Service
@@ -25,14 +29,13 @@ public class FoodMenuService {
 	
 	public ResponseEntity<ResponseStructure<FoodMenu>> saveFoodMenu(FoodMenu foodMenu, long userId)
 	{
-		User user= userDao.findById(userId);
-		if(user!=null)
+		Optional<User> user= userDao.findById(userId);
+		if(user.isPresent())
 		{
-			Role role=user.getRole();
+			Role role=user.get().getRole();
 			if(role.equals(Role.ADMIN) || role.equals(Role.MANAGER))
 			{
 				FoodMenu menu=foodMenuDao.saveFoodMenu(foodMenu);
-				if(menu!=null) {
 				ResponseStructure<FoodMenu> responseStructure=new ResponseStructure<>();
 				responseStructure.setStatus(HttpStatus.CREATED.value());
 				responseStructure.setMessage("Menu Data saved sucessfully");
@@ -41,25 +44,22 @@ public class FoodMenuService {
 				ResponseEntity<ResponseStructure<FoodMenu>> responseEntity=new ResponseEntity<ResponseStructure<FoodMenu>>(responseStructure,HttpStatus.CREATED);
 
 				return responseEntity;
-				}
-				else
-					throw new NoSuchDataFoundException();
 			}
 			else
-				throw new NoSuchDataFoundException();
+				throw new UserRoleDoesNotMatch();
 		}
 		else
-			throw new NoSuchDataFoundException();
+			throw new UserDataNotFoundInTheDatabase();
 	}
 	
 	public ResponseEntity<ResponseStructure<FoodMenu>> findFoodMenuById(Long id)
 	{
-		FoodMenu menu=foodMenuDao.findFoodMenuById(id);
-		if(menu!=null) {
+		Optional<FoodMenu> menu=foodMenuDao.findFoodMenuById(id);
+		if(menu.isPresent()) {
 		ResponseStructure<FoodMenu> responseStructure=new ResponseStructure<>();
 		responseStructure.setStatus(HttpStatus.OK.value());
 		responseStructure.setMessage("Menu Data retrived sucessfully");
-		responseStructure.setData(menu);
+		responseStructure.setData(menu.get());
 		
 		ResponseEntity<ResponseStructure<FoodMenu>> responseEntity=new ResponseEntity<ResponseStructure<FoodMenu>>(responseStructure,HttpStatus.OK);
 
@@ -68,6 +68,8 @@ public class FoodMenuService {
 		else
 			throw new NoSuchDataFoundException();
 	}
+	
+	
 	public ResponseEntity<ResponseStructure<List<FoodMenu>>> findAllFoodMenu()
 	{
 		List<FoodMenu> menu=foodMenuDao.findAllFoodMenu();
@@ -87,19 +89,46 @@ public class FoodMenuService {
 		
 	public ResponseEntity<ResponseStructure<String>> removeById(Long id)
 	{
-		String str=foodMenuDao.removeFoodMenuById(id);
-	if(str!=null)
-	{
-		ResponseStructure<String>  responseStructure=new ResponseStructure<>();
-		responseStructure.setStatus(HttpStatus.OK.value());
-		responseStructure.setMessage("Menu removed Sucessfully");
-		responseStructure.setData(str);
+		Optional<FoodMenu> foodMenu=foodMenuDao.findFoodMenuById(id);
 		
-		ResponseEntity<ResponseStructure<String>> responseEntity=new ResponseEntity<ResponseStructure<String>>(responseStructure, HttpStatus.OK);
-		return responseEntity;
+		if(foodMenu.isPresent()) {
+					String str=foodMenuDao.removeFoodMenuById(id);
+				if(str!=null)
+				{
+					ResponseStructure<String>  responseStructure=new ResponseStructure<>();
+					responseStructure.setStatus(HttpStatus.NO_CONTENT.value());
+					responseStructure.setMessage("Menu removed Sucessfully");
+					responseStructure.setData(str);
+					
+					ResponseEntity<ResponseStructure<String>> responseEntity=new ResponseEntity<ResponseStructure<String>>(responseStructure, HttpStatus.NO_CONTENT);
+					return responseEntity;
+				}
+				else
+					throw new NoSuchDataFoundException();
+				}
+		else {
+			throw new NoSuchDataFoundException();
+		}
 	}
-	else
-		throw new NoSuchDataFoundException();
+	
+	public ResponseEntity<ResponseStructure<FoodMenu>> updateFoodMenu(long foodMenuId,FoodMenu foodMenu)
+	{
+		Optional<FoodMenu> optional=foodMenuDao.findFoodMenuById(foodMenuId);
+		
+		if(optional.isPresent())
+		{
+			foodMenu.setId(foodMenuId);
+			ResponseStructure<FoodMenu>  responseStructure=new ResponseStructure<>();
+			responseStructure.setStatus(HttpStatus.OK.value());
+			responseStructure.setMessage("Menu Updated Sucessfully");
+			responseStructure.setData(foodMenu);
+			
+			ResponseEntity<ResponseStructure<FoodMenu>> responseEntity=new ResponseEntity<ResponseStructure<FoodMenu>>(responseStructure, HttpStatus.OK);
+			return responseEntity;
+	
+		}
+		else
+			throw new FoodMenuIdIsNotPresent();
 	}
 
 }
